Added a "get-help" shortcut for PowerShell!
In order to get help for a command, you could press {Home}, type in "get-help" (whether you press {TAB} or not to complete it), add a space, press {End} to go all the way back to the end of the command, and then add another space. With this new AHK function, it all does so by the press of a hotkey (or keyboard shortcut) within just a second!